+++ templates/shared_services/cyclecloud/terraform/data.tf
@@ -1,17 +1,3 @@
-data "azurerm_subnet" "shared" {
-  resource_group_name  = local.core_resource_group_name
-  virtual_network_name = local.core_vnet
-  name                 = "SharedSubnet"
-}
-data "azurerm_key_vault" "keyvault" {
-  name                = local.keyvault_name
-  resource_group_name = local.core_resource_group_name
-}
-
-data "azurerm_resource_group" "rg" {
-  name = local.core_resource_group_name
-}
-
 data "azurerm_key_vault_key" "tre_encryption_key" {
   count        = var.enable_cmk_encryption ? 1 : 0
   name         = local.cmk_name